Started styling teamList page
