--- front-end/sentimentbot/src/Components/teamlist/teamlist.css
@@ -1,5 +1,0 @@
-.teamlist-container {
-    width: 20%;
-    border: 1px solid black;
-    margin: 0 auto;
-}
