+++ front-end/sentimentbot/src/Components/teamlistTest/teamlist 2.css
